Refatoração dos codigos e criação de metodos genericos
